[JENKINS-40929_Developer_can_see_durations_when_hovering_on_indicators] move component to core to be able to reuse it in personalization
