+++ blueocean-core-js/src/js/components/LiveStatusIndicator.jsx
@@ -1,17 +1,15 @@
 import React, { Component, PropTypes } from 'react';
 import { LiveStatusIndicator as LiveStatusIndicatorJdl } from '@jenkins-cd/design-language';
-import { logging } from '@jenkins-cd/blueocean-core-js';
-import { TimeHarmonizer as timeHarmonizer } from './TimeHarmonizer';
-const logger = logging.logger('io.jenkins.blueocean.dashboard.LiveStatusIndicator');
+import { TimeHarmonizer as timeHarmonizer } from '../index';
+import logging from '../logging';
+const logger = logging.logger('io.jenkins.blueocean.core.LiveStatusIndicator');
 
 export class LiveStatusIndicator extends Component {
 
     render() {
         const { result, getI18nTitle } = this.props;
         const title = getI18nTitle(result);
-//      const title = result;
         logger.debug('i18n title', title);
-
         return (
             <div title={title}>
                 <LiveStatusIndicatorJdl { ... this.props } />
