+++ blueocean-core-js/src/js/index.js
@@ -93,3 +93,8 @@ export const DEBUG = {
     enableMocksForI18n,
     disableMocksForI18n,
 };
+
+export { TimeManager } from './utils/serverBrowserTimeHarmonize';
+
+export { TimeHarmonizer } from './components/TimeHarmonizer';
+export { LiveStatusIndicator } from './components/LiveStatusIndicator';
