+++ blueocean-dashboard/src/main/js/components/Branches.jsx
@@ -4,10 +4,9 @@ import {
     ReadableDate,
     WeatherIcon,
 } from '@jenkins-cd/design-language';
-import { RunButton } from '@jenkins-cd/blueocean-core-js';
+import { RunButton, LiveStatusIndicator } from '@jenkins-cd/blueocean-core-js';
 import Extensions from '@jenkins-cd/js-extensions';
 import { observer } from 'mobx-react';
-import LiveStatusIndicator from './LiveStatusIndicator';
 import { CellRow, CellLink } from './CellLink';
 
 import { buildRunDetailsUrl } from '../util/UrlUtils';
